+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## Version 2.0.0 - Released 2020.Oct.25
+## Version 5.6.1 - Released 2021.Jan.24
+- Fix for submit button
+
+## Version 5.6.0 - Released 2021.Jan.24
+- Fixes classnames for playlists
+
+## Version 5.5.2 - Released 2020.Dec.20
+- Fixes missing fields like displayannotations and shorttitle
+
+## Version 5.5.1 - Released 2020.Dec.20
+- Update to fix scripts in embedded json
+
+## Version 5.5.0 - Released 2020.Oct.25
 - Adds new embed options to plugin including audio, dropboxes, playlists, and quizzes.
 - Adds css codes that can be overridden in WordPress